PC 8250 Driver Interrupts Finally Working
There were a number of issues with my implementation of the 8250 driver,
one of the big one's was that interrupts were not working in general, so
I added the "NO_WAIT" option to device flags to avoid the problem and
still have a "working" driver.

Also, this has only been tested on single core qemu so far, so I'm not
sure how SMP will handle it, especially with the changes I made to
registering the LAPIC as a "percpu" interrupt device.