Mapping RISCV cpu_id_t to HARTID and vice versa
There's no promise that RISCV hartid's will be contiguous (there is a
promise of a hart 0, but it could theoretically be hart 0 and hart 2,
and not hart 0 and hart 1).

Various parts of the kernel depend on a contiguous range of cpu_id_t
from [0, nk_num_cpus()), so we need to maintain a mapping between the
two different ranges of cpu identifiers.