Consistent "make clean" Rule
Forces the inclusion of "clean.mk" files found in subdirectories of the
scripts directory into the root Makefile. This makes it so CLEAN_RULES
can be the same regardless of config. Otherwise, if you build once and
swap your config, you may not get rid of all the generated files.

Ex. Previously if you built for ARM64, you'll end up with uImage, but if you swap
to x64 and ran "make clean" it wouldn't delete the uImage file.

Also moved qemu and objdump rules into their own files to declutter the
root Makefile