Simpler Root Makefile, and handling CXX files
Some of the logic of "scripts/build.mk" was essentially being replicated
in the root Makefile, which wasn't very easy to read. Now instead,
build.mk is invoked on the root nautilus directory and the Makefile
provides obj-y like all the Makefiles in the src directory.

Also, a minor addition to re-add compiling .cc files correctly for NAUT_CONFIG_CXX_SUPPORT. We still need to support libstdc++.a correctly though.