Chardev Console Copy from Buffer Fix on x86
Copying the buffer from a vc to the chardev console when switching VC's
was printing all lines on top of one another on x86. This fixes that.

(This can cause ARM64 to have extra blank lines in the same situation
sometimes, but I'd rather ARM have extra lines than x86 not work at all)