Added MSI Funtions to nk_irq_dev interface
msi_addr() and msi_x_addr() should provide the address a device needs to
write to for a given IRQ marked with the NK_IRQ_DESC_FLAG_MSI(_X) flag.

msi_msg() and msi_x_msg() should give the specific data to be written to
that address for the device.

msi_block_size() should give the number of "contiguous" interrupts
starting at the specific hwirq, (where 2^N contiguous interrupts have the
same msi_msg() data except for the lowest N bits, and the "base" hwirq
which is handed to msi_block_size() has all lowest N bits as zero)
On x86 this just corresponds to the largest (up to 32) power of 2 range of
interrupts which is aligned to it's own size starting on the base IRQ.

msi_index_block() takes a "base" hwirq as in msi_block_size() and
returns the hwirq of the "n-th" interrupt in the block. This is 100%
overkill for now but just serves to futureproof, and avoid the kernel
making any assumptions whatsoever about how nk_hwirq_t are arranged. On
x86 this is just simple addition, base_irq + index.