Split x86 init function into two parts
The single most "prone to breaking" part of the kernel I've come across
before is the call to smp_ap_stack_switch halfway through init, because
the compiler really fights you on use of the stack.

So instead of trying to keep the same call frame going while changing
stacks, instead we start in "boot_stack_init" until needing to change
stacks, and then return from the function, dropping out of a C
environment entirely, change stacks, and then call an entirely new
function "threaded_init" with it's own stack frame on the new
dynamically allocated stack.

The same trick is then used for the AP initialization too, and all of
this just makes the whole init process a lot less prone to suddenly
breaking for hard to diagnose stack related issues.