Starting to implement "init_stages"
The module system wasn't scaling well for handling dependencies,
so we're going with the simpler solution of having a set of "init"
functions which can defer their own execution to be run again later if a
dependency isn't up yet (Very similarly to Linux initcalls and
EPROBE_DEFER).

The set of stages is controlled through adding X macros to the list in
include/nautilus/init_stages.h, and the order of execution within a
stage is never promised.