8250 Serial Driver Fixes
Fixed race condition in dev/8250/core.c which could occur because the
waitqueue lock and the output lock on the UART were acquired in a
different order in two different places. (Fix is in
uart_8250_kick_output).

Also dealt with an issue in the dw_8250 UART driver where receive
interrupts would be ignored entirely because the core.c irq_handler
wasn't always being invoked.

Besides that just some additional checks and warnings when enabling
FIFO's (added when debugging, but seem useful enough to leave in place).