Scheduler Printing Formatting
Reformatted some scheduler printing statements to provide more info / be
easier for people who are unfamiliar with nautilus' scheduler to read.

The most noticeable change is to the printout when calling "threads" in
the shell, it used to use a large number of post-fixes to indicate what
all of the numbers are, but now it's on multiple lines and indicates
what each number is exactly. It would be nice to make it so arguments
to "threads" can reduce the amount of printing though, especially for a
large number of threads appearing on VGA (I'm willing to be that's why
it was so condensed and on a single line beforehand).

Also, there's no need for the hack to "preallocate" space on the stack
to avoid the compiler messing up on stack switches, because of the split
init functions now (still need to patch ARM64 to use split init though).