Slightly larger stack sizes for O0
Stack overruns are one of the most common issues when running on O0, so
to help deal with this, the stack sizes of some common threads are
increased here. They've also been moved to macro's so that they could be
easily modified / set depending on the value of
NAUT_CONFIG_COMPILER_OPT_LEVEL or something like that.