Fixing x86 Interrupt Bug
This bug has been in the kernel since commit
(d11aa03ec6aa5bdfbe2709ae3fe32e76d638702f) when it was first introduced.

It resulted in no interrupts being delivered (or at least very rarely
delivered) on x64. From what I can tell, its due to the behaviour of the
"sti" instruction where it will delay taking effect for a single
instruction afterwards, and so there were very specific situations where
it wouldn't actually allow interrupts to be enabled? I'm not 100% sure
on this one.

I'm 100% sure it was introduced in the commit mentioned above (from
bisecting). And that if you revert the file "include/nautilus/cpu_state.h" then the issue goes away in the commit.
And after playing around with different combinations of the changes in
the file, swapping sti() and arch_enable_ints(), was the culprit.
So because arch_enabled_ints() isn't inlined while sti() was, I'm
guessing the problem is because of the delayed effect of the sti
instruction. (And this is backed up by the fact introducing a "nop"
after the "sti" instruction seems to have made the issue go away).