Update sincos Shell Command
Made the sincos command take an argument (still has to be an integer
because sscanf doesn't seem to work with %f). Also made the actual
command "sincos" like the help message, not "sc".

There are two copies of this command still becasue there are two copies
of all the math function (one for libm and one for NAS) which is a
bigger issue which I'll get around to addressing later.