x64 Stack Alignment and QEMU Bug
The "nop" after the "sti" instruction in "arch_enable_ints" seems to
only be necessary when running on QEMU (pressumably because of the TCG),
but on real hardware it's been tested and functions correctly.

On QEMU without the nop the idle loop never receives interrupts if
"HALT_WHILE_IDLE" is disabled, so for now we'll just leave that as the
default setting and consider just enforcing the halt to make QEMU work
correctly with the TCG running.