ALLOW_UNINTEGRATED Kconfig Option
Integration has become very very difficult with how large and entangled
the whole kernel is. At this point I can get a working kernel with a bit
of effort if I'm willing to go in and make tweaks based on whatever
configuration "do you want a RISC-V machine with 6 smp cores? Sure give
me a few hours and I (who has lots of experience with this specific
kernel) can go get it working for you (and in the meantime break the x64
APIC driver in new and horrifying ways)".

So to help deal with this I'm putting down the heavy hammer and putting
almost all kernel features behind the "ALLOW_UNINTEGRATED" flag.

This way the smallest subset of features which are needed for the kernel
can be tested and bugfixed (all together) and ideally add some basic
configuration fuzzing (without worrying that a half decade old class
project will the the source of an error).