NAUT_CONFIG_USE_NAUT_BUILTINS was incorrect
The __builtin_* string functions do not actually provide implementations
of the symbols. The compiler will just optimize their usage if it can
determine the result from the inputs.

For example the compile does not usually optimize `strlen("hello")` but it CAN
optimize `__builtin_strlen("hello")` by replacing it with `5`.
After optimization, the compiler then replaces all `__builtin_*` string
functions which did not get removed by optimizations and replaces them
with calls to the original functions. So even if we are using the
`__builtin_*` versions, Nautilus must include it's own implementations
of all string functions as well.